// src/api/endpoints/model/testRecord.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * (title)
 * OpenAPI spec version: 0.0.0
 */

export interface TestRecord {
  id: string;
  tenantId: string;
  sampleId: string;
  parameterCode: string;
  standardCode?: string;
  requirementCode?: string;
  requirement: string;
  result: string;
  verdict?: string;
  createdAt: string;
  updatedAt: string;
}